feat: service class refactoring
